new camera orientation algorithm by https://github.com/Limint
