+++ BetterGenshinImpact/BetterGenshinImpact.csproj
@@ -10,7 +10,7 @@
 		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
 		<ApplicationIcon>Assets\Images\logo.ico</ApplicationIcon>
 		<AssemblyName>BetterGI</AssemblyName>
-		<AssemblyVersion>0.37.2</AssemblyVersion>
+		<AssemblyVersion>0.37.3</AssemblyVersion>
 		<Platforms>x64</Platforms>
 		<DebugType>embedded</DebugType>
 	</PropertyGroup>
@@ -48,7 +48,6 @@
 		<PackageReference Include="Microsoft.ML.OnnxRuntime.DirectML" Version="1.18.1" />
 		<PackageReference Include="Microsoft.ML.OnnxRuntime.Managed" Version="1.18.1" />
 		<PackageReference Include="Microsoft.Web.WebView2" Version="1.0.2592.51" />
-		<PackageReference Include="NumpyDotNet" Version="0.9.86.2" />
 		<PackageReference Include="Ookii.Dialogs.Wpf" Version="5.0.1" />
 		<PackageReference Include="OpenCvSharp4.WpfExtensions" Version="4.8.0.20230708" />
 		<PackageReference Include="OpenCvSharp4.Extensions" Version="4.8.0.20230708" />
